extract file contents view into its own file
